refactor(ALL) : MVC 디자인 패턴에 맞게끔 구조 재배치
